Optimize HTMLConstructionSite::mergeAttributesFromTokenIntoElement() https://bugs.webkit.org/show_bug.cgi?id=244604
Reviewed by NOBODY (OOPS!).

Avoid duplicate calls to findAttributeByName() for each attribute both inside
mergeAttributesFromTokenIntoElement() and Element::setAttribute(). Instead,
introduce a new Element::setAttributeWithoutOverwriting() and drop now
unnecessary call to findAttributeByName() at call site.

This is a 0.6% progression on Speedometer on Intel and seems to be a 0.3%
progression on Apple Silicon.

* Source/WebCore/dom/Element.cpp:
(WebCore::Element::setAttributeWithoutOverwriting):
* Source/WebCore/dom/Element.h:
* Source/WebCore/html/parser/HTMLConstructionSite.cpp:
(WebCore::HTMLConstructionSite::mergeAttributesFromTokenIntoElement):