Test addition (248077@main): [ macOS wk1 Debug ] http/tests/loading/form-submission-no-load-event.html is a flaky crash https://bugs.webkit.org/show_bug.cgi?id=244611 <rdar://99386658>
Reviewed by NOBODY (OOPS!).

Submitting the form schedules an async task to do the navigation in a new window.
We need to make sure that we don't call `testRunner.notifyDone()` before this
navigation occurs. Otherwise, we will hit an assertion in DRT indicating we're
trying to open a new window but waitUntilDone is not (no longer) true.

* LayoutTests/http/tests/loading/form-submission-no-load-event-expected.txt:
* LayoutTests/http/tests/loading/form-submission-no-load-event.html:
* LayoutTests/platform/mac-wk1/TestExpectations: